requirements.txt
python-dotenv==0.19.0
pymongo==3.12.0
PyYAML==5.4.1
Flask==2.0.1
requests-html==0.10.0
flask_swagger_ui==3.36.0
flask_cors==3.0.10